% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{assign_map}
\alias{assign_map}
\title{Assign edges to the hex grid}
\usage{
assign_map(
  proj,
  eccentricity = 0.9,
  assign_type = 1,
  report_progress = TRUE,
  pb_markdown = FALSE
)
}
\arguments{
\item{proj}{object of class \code{pm_project}.}

\item{eccentricity}{eccentricity of ellipses, defined as half the distance
between foci divided by the semi-major axis. \eqn{e = sqrt{1 - b^2/a^2}},
where \eqn{e} is the eccentricity, \eqn{a} is the length of the semi-major
axis, and \eqn{b} is the length of the semi-minor axis. Eccentricity ranges
between 0 (perfect circle) and 1 (straight line between foci).}

\item{assign_type}{Assignment method to use (1=standard, 2=new method with 
function for dealing with duplicate ellipses)}

\item{report_progress}{if \code{TRUE} then a progress bar is printed to the
console during the permutation testing procedure.}

\item{pb_markdown}{whether to run progress bars in markdown mode, in which
case they are updated once at the end to avoid large amounts of output.}
}
\description{
Given an PlasmoMAPI project with a hex map already created, determine
  which edges intersect each hex. Assumes an elliptical projection with the
  start- and end-points of an edge becoming the two foci of an ellipse.
}
