% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{calc_simple_hex_values}
\alias{calc_simple_hex_values}
\title{Calculate map hex values without permutation}
\usage{
calc_simple_hex_values(proj)
}
\arguments{
\item{proj}{object of class \code{pm_project}.}
}
\description{
TODO
}
