% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{check_PlasmoMAPI_loaded}
\alias{check_PlasmoMAPI_loaded}
\title{Check that PlasmoMAPI package has loaded successfully}
\usage{
check_PlasmoMAPI_loaded()
}
\description{
Simple function to check that PlasmoMAPI package has loaded
  successfully. Prints "PlasmoMAPI loaded successfully!" if so.
}
