% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_significant_hexes}
\alias{get_significant_hexes}
\title{Get significant hexes}
\usage{
get_significant_hexes(
  proj,
  empirical_tail = "both",
  FDR = 0.05,
  min_hex_coverage = 10
)
}
\arguments{
\item{proj}{object of class \code{pm_project}.}

\item{empirical_tail}{whether to calculate empirical p-values using a
one-sided test (\code{empirical_tail = "left"} or \code{empirical_tail =
"right"}) or a two-sided test (\code{empirical_tail = "both"}).}

\item{FDR}{the false discovery rate, i.e. the probability that a hex
identified as significant is actually a false positive.}

\item{min_hex_coverage}{minimum coverage (number of edges assigned to a hex)
for it to be included in the final result.}
}
\description{
Given a completed PlasmoMAPI analysis, return a list of which
  hexes are significant outliers.
}
