% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pm_project.R
\name{is.pm_project}
\alias{is.pm_project}
\title{Determine if object is of class pm_project}
\usage{
is.pm_project(x)
}
\arguments{
\item{x}{object to query.}
}
\description{
Determine if object is of class \code{pm_project}.
}
