% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_coverage}
\alias{plot_coverage}
\title{Plot hex coverage}
\usage{
plot_coverage(proj, breaks = c(0, 10, 20, 30, 40, 50, 100, Inf))
}
\arguments{
\item{proj}{object of class \code{pm_project}.}

\item{breaks}{the sequence of coverage breaks used.}
}
\description{
Given an PlasmoMAPI project with a map already assigned, plot
  the coverage (number of intersecting ellipses) of each hex. This plot acts
  as a diagnostic, as hexes with low coverage will conform less well to the
  assumptions of the permutation test (see \emph{details}).
}
\details{
Good coverage is needed to ensure the validity of the statistical
  procedure. For each hex, the observed value is the mean of the normalised
  values of the edges that intersect it. Under the null model that these
  normalised values have no systematic bias, the distribution of this test
  statistic is approximately normal as a consequence of the central limit
  theorem. Hence, we can use a permutation test to characterise the mean and
  standard deviation of this null distribution, then we can quantify how
  extreme the observed value is in terms of its z-score. However, if coverage
  is too low then the null distribution will not be normally distributed, and
  hence the z-score will not be an accurate description of how extreme the
  observed data really is.
}
