% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_ellipses}
\alias{plot_ellipses}
\title{Plot network of nodes and ellipses}
\usage{
plot_ellipses(
  proj,
  eccentricity = 0.9,
  n = 30,
  alpha = 0.4,
  col_scale = rev(PlasmoMAPI::col_hotcold()),
  zlim = NULL,
  node_size = 2,
  base_plot = NULL
)
}
\arguments{
\item{proj}{object of class \code{pm_project}.}

\item{eccentricity}{eccentricity of ellipses, defined as half the distance
between foci divided by the semi-major axis. \eqn{e = sqrt{1 - b^2/a^2}},
where \eqn{e} is the eccentricity, \eqn{a} is the length of the semi-major
axis, and \eqn{b} is the length of the semi-minor axis. Eccentricity ranges
between 0 (perfect circle) and 1 (straight line between foci).}

\item{n}{number of points that make up each ellipse.}

\item{alpha}{the opacity of ellipses.}

\item{col_scale}{the colour scale to use.}

\item{zlim}{the limits of the colour scale. If \code{NULL} then these limits
are chosen automatically.}

\item{node_size}{the plotted size of nodes.}

\item{base_plot}{optional base plot (object of class \code{ggplot}) on which
this function builds. If \code{NULL} then a simple empty plot is used.}
}
\description{
Produce a simple plot of nodes connected by all pairwise
  ellipses, where nodes make up the foci of ellipses and colours correspond
  to statistical distances. Edges with \code{NA} values are not plotted.
}
