% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_map}
\alias{plot_map}
\title{Plot hex map of PlasmoMAPI output}
\usage{
plot_map(
  proj,
  col_scale = viridisLite::magma(100),
  plot_sampling_points = TRUE,
  plot_hex_grid = TRUE,
  min_hex_coverage = 10,
  plot_significance = TRUE,
  empirical_tail = "both",
  alpha_raw = 0.05,
  base_plot = NULL,
  poly_list = list()
)
}
\arguments{
\item{proj}{object of class \code{pm_project}.}

\item{col_scale}{the colour scale to use.}

\item{plot_sampling_points}{whether to overlay sampling locations.}

\item{plot_hex_grid}{whether to plot the hex map. If the project contains
output then hexes will be coloured based on z-scores, otherwise a flat
colour scheme will be used.}

\item{min_hex_coverage}{minimum coverage (number of edges assigned to a hex)
for it to be included in the final result, otherwise these hexes are given
the value \code{NA}.}

\item{plot_significance}{whether to outline areas that were identified as
significant outliers.}

\item{empirical_tail}{if plotting significance, whether to calculate
empirical p-values using a one-sided test (\code{empirical_tail = "left"}
or \code{empirical_tail = "right"}) or a two-sided test
(\code{empirical_tail = "both"}).}

\item{alpha_raw}{the significance threshold used to determine significantly
high/low values. This raw value is Bonferroni corrected based on the
effective number of independent samples, and hence applies to the whole map
and not just a single hex.}

\item{base_plot}{optional base plot (object of class \code{ggplot}) on which
this function builds. If \code{NULL} then a simple empty plot is used.}

\item{poly_list}{optional list of polygon coordinates that are added to plot.}
}
\description{
Plot hex map of PlasmoMAPI output.
}
