% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_network}
\alias{plot_network}
\title{Plot network of nodes and edges}
\usage{
plot_network(
  proj,
  col_scale = rev(PlasmoMAPI::col_hotcold()),
  zlim = NULL,
  node_size = 2,
  edge_size = 1,
  base_plot = NULL
)
}
\arguments{
\item{proj}{object of class \code{pm_project}.}

\item{col_scale}{the colour scale to use.}

\item{zlim}{the limits of the colour scale. If \code{NULL} then these limits
are chosen automatically.}

\item{node_size, edge_size}{the plotted size of nodes and edges.}

\item{base_plot}{optional base plot (object of class \code{ggplot}) on which
this function builds. If \code{NULL} then a simple empty plot is used.}
}
\description{
Produce a simple plot of nodes connected by all pairwise edges,
  where edge colours correspond to statistical distances. Edges with
  \code{NA} values are not plotted.
}
