% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{pm_analysis}
\alias{pm_analysis}
\title{Perform PlasmoMAPI analysis}
\usage{
pm_analysis(
  proj,
  n_perms = 1000,
  n_breaks = 50,
  min_dist = 0,
  max_dist = Inf,
  min_group_size = 5,
  aggregate_clusters = FALSE,
  report_progress = TRUE,
  pb_markdown = FALSE
)
}
\arguments{
\item{proj}{object of class \code{pm_project}.}

\item{n_perms}{number of permutations in test.}

\item{n_breaks}{values are broken into this many groups based on spatial
distances. Permutation testing then occurs within groups. Hence, a larger
value of \code{n_breaks} does a better job of conditioning on spatial
distance, but comes at the cost of statistical power in permutation
testing. A warning is printed if any group ends up containing fewer than 10
edges.}

\item{min_dist, max_dist}{minimum and maximum edge lengths to be included in
the analysis. Anything outside this range is ignored.}

\item{min_group_size}{minimum number of edges within a spatial permutation
group, otherwise edges within this group are replaced with \code{NA}.}

\item{aggregate_clusters}{if \code{TRUE} then observations from the same
sampling location are aggregated prior to running the analysis. The
statistical value between locations X and Y becomes the mean of all
pairwise values sampled at X and Y. If \code{FALSE} then each value is
considered separately.}

\item{report_progress}{if \code{TRUE} then a progress bar is printed to the
console during the permutation testing procedure.}

\item{pb_markdown}{whether to run progress bars in markdown mode, in which
case they are updated once at the end to avoid large amounts of output.}
}
\description{
TODO.
}
